requirements.txt
amqp==5.0.9
asgiref==3.4.1
astroid==2.9.3
billiard==3.6.4.0
boto3==1.20.34
botocore==1.23.34
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.1
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
django-mysql==4.3.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-recursive==0.1.2
isort==5.10.1
jmespath==0.10.0
kombu==5.2.3
lazy-object-proxy==1.7.1
mccabe==0.6.1
minio==7.1.2
numpy==1.22.0
pandas==1.3.5
Pillow==9.0.0
platformdirs==2.4.1
prompt-toolkit==3.0.24
pycparser==2.21
PyJWT==1.7.1
pylint==2.12.2
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.29
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.8
uuid==1.30
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
yapf==0.32.0

channels~=3.0.4